+++ pmd-core/src/main/java/net/sourceforge/pmd/cpd/CPDConfiguration.java
@@ -65,6 +65,7 @@ public class CPDConfiguration extends AbstractConfiguration {
 
     private boolean ignoreIdentifierAndLiteralSequences = false;
 
+    @Deprecated
     private boolean skipLexicalErrors = false;
 
     private boolean noSkipBlocks = false;
@@ -226,10 +227,20 @@ public class CPDConfiguration extends AbstractConfiguration {
         this.ignoreIdentifierAndLiteralSequences = ignoreIdentifierAndLiteralSequences;
     }
 
+    /**
+     * @deprecated This option will be removed. With {@link #isFailOnError()}, you can
+     * control whether lexical errors should fail the build or not.
+     */
+    @Deprecated
     public boolean isSkipLexicalErrors() {
         return skipLexicalErrors;
     }
 
+    /**
+     * @deprecated This option will be removed. With {@link #setFailOnError(boolean)}, you can
+     * control whether lexical errors should fail the build or not.
+     */
+    @Deprecated
     public void setSkipLexicalErrors(boolean skipLexicalErrors) {
         this.skipLexicalErrors = skipLexicalErrors;
     }